making json use uncode strings exclusively.
Also, a few other minor json-related fixes.

This fixes auto-ported test_json for python2 except for athena-related
failures.

And yes, it's questionable to auto-decode byte strings as utf-8 in
parse; but (a) other failures aren't much prettier and (b) if you want
to be sure, just decode yourself and pass a unicode string.